Revert vid field change and make new type instead
In order to handle potentially tagged traffic, a new struct and function
were created: `VlanEthernetFrame` and `parse_vlan_ethernet_frame`. This
is to keep the public interface of the `EthernetFrame` struct the same
and remove the overhead of parsing for potentially tagged traffic in the
common use case where traffic is guaranteed to be untagged.